refactor: link-style, link-style--hover-effect class
