+++ src/components/App/Header/index.tsx
@@ -1,11 +1,9 @@
 import { memo } from 'react';
 
 import { Container, getSize } from '@/components/Functional/Container';
-import { Anchor as _Anchor } from '@/components/UI/Anchor';
-import { LinkStyle } from '@/components/UI/LinkMenu';
 import { Logo } from '@/components/UI/Logo';
 import { useSearch } from '@/components/UI/Search';
-import { styled } from '@/ui/styled';
+import { styled } from '@/ui/styled/dynamic';
 import { HeaderLayout } from './HeaderLayout';
 
 export default memo(function Header() {
@@ -32,13 +30,4 @@ const HeaderContainer = styled(Container)`
   justify-content: space-between;
   height: 100%;
   margin: 0 auto;
-`;
-
-export const Anchor = styled(_Anchor)`
-  display: flex;
-  align-items: center;
-  height: 70%;
-  pointer-events: auto;
-
-  ${LinkStyle}
 `;
