+++ src/components/UI/Logo/index.tsx
@@ -1,19 +1,16 @@
 import { Anchor as _Anchor } from '@/components/UI/Anchor';
-import { LinkStyle } from '@/components/UI/LinkMenu';
 import { SITE_NAME } from '@/constant';
-import { styled } from '@/ui/styled';
+import { styled } from '@/ui/styled/dynamic';
 
 export const Logo = () => {
   return (
-    <Anchor href="/" prefetch={false}>
+    <Anchor className="link-style link-style--hover-effect" href="/" prefetch={false}>
       <img src="/logo.v2.svg" alt={SITE_NAME} height="25" width="80" />
     </Anchor>
   );
 };
 
 const Anchor = styled(_Anchor)`
-  ${LinkStyle}
-
   display: flex;
   align-items: center;
   padding: var(--space-1);
