+++ src/pages/_components/home/LinkMore/index.tsx
@@ -1,13 +1,12 @@
 import type { ReactNode } from 'react';
 
 import { Anchor as _Anchor } from '@/components/UI/Anchor';
-import { LinkStyle } from '@/components/UI/LinkMenu';
 import { CaretRightIcon, ICON_SIZE_XS } from '@/ui/icons';
-import { styled } from '@/ui/styled';
+import { styled } from '@/ui/styled/dynamic';
 
 export function LinkMore({ href, text }: { href: string; text: string | ReactNode }) {
   return (
-    <Anchor href={href}>
+    <Anchor className="link-style link-style--hover-effect" href={href}>
       {text}
       <CaretRightIcon aria-hidden="true" fill="currentColor" height={ICON_SIZE_XS} width={ICON_SIZE_XS} />
     </Anchor>
@@ -15,7 +14,5 @@ export function LinkMore({ href, text }: { href: string; text: string | ReactNod
 }
 
 const Anchor = styled(_Anchor)`
-  ${LinkStyle}
-
   font-size: var(--font-size-sm);
 `;
